feat(renderer, docs): complete P5 cross-platform long-term acceptance
1. 重构滚动回退reflow逻辑：live-bottom模式下取消自动全历史reflow，仅在用户进入回看时惰性生成布局
2. 新增渲染统计指标：atlas内存、光栅队列状态、滚动回退reflow请求计数等
3. 更新架构文档与P5验收说明，补充Windows D3D/DPR/长稳验收细节
4. 重构P5 benchmark，新增多配置参数、自动收敛等待与严格验收规则
5. 修复TerminalCore.cpp中属性类型转换隐式警告

+++ src/core/terminal/TerminalCore.cpp
@@ -78,7 +78,7 @@ quint64 rowContentIdentity(const NovaTerm::Cell* cells, int columns)
         mix(cell.background.red | (cell.background.green << 8)
             | (cell.background.blue << 16));
         const auto& a = cell.attributes;
-        quint64 attributes = a.bold
+        quint64 attributes = quint64(a.bold)
             | (quint64(a.underline) << 1)
             | (quint64(a.italic) << 2)
             | (quint64(a.blink) << 3)
